~ Fixes case bug in field name produced by node code generator. + Adds new field named 'lookAt' in Camera node to specify position and orientation. ~ Renames Camera field 'matrix' into 'projection'. Old accessors are always provided, but are deprecated. ~ Updates nodes generated from an xml description. ~ Updates BasicViewer class to use 'lookAt' field of Camera node. vgUI::BasicViewer::getViewTransformation() method is deprecated.

+++ trunk/engines/vgeGL/src/handler/painter/VertexShape.cpp
@@ -114,7 +114,7 @@ void VertexShape::apply( vge::engine::Engine *pEngine, vgd::node::Node *pNode )
 						vgd::Shp< GLSLState::TexUnitState > texUnitState = glslState.getTexture( unit );
 						if ( texUnitState )
 						{
-							texUnitState->setTexCoordDim( dimTexCoord );
+							texUnitState->setTexCoordDim( static_cast< uint8 >(dimTexCoord) );
 						}
 						// else nothing to do
 					}
